layout: extract the shared geometry kernel to src/layout/geometry.ts (step 2, part 1)
Step 2 precursor (spec OQ-D1): move the leaf geometry helpers the post-ELK
passes depend on into a leaf module — DEFAULTS, flattenGroupBounds, the
polyline/path math (polylineLength, pointAtPathDistance, calculatePathMidpoint),
and the direction/axis helpers (layoutFlow, positionedNodeCenter,
nodeMainStart/CrossStart/CrossSize/MainSize/CrossCenter/MainCenter, rectsOverlap).
layout-engine.ts now imports them.

This is the dependency-direction precursor: once the passes move to
src/layout/passes/, they import these from geometry.ts and never from
layout-engine.ts, keeping the graph one-directional (layout-engine -> passes ->
geometry) and cycle-free.

Pure relocation, zero behavior change. Verified byte-identical: determinism
oracle 370 pass / 0 fail / 52 snapshots; full unit suite 4147 pass / 0 fail
across 183 files; typecheck clean.
